REFACTOR `Utils` namespace to `Services`
